+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -9,6 +9,8 @@ import * as runtimeEnv from '@expo/env';
 import { SerialAsset } from '@expo/metro-config/build/serializer/serializerAssets';
 import assert from 'assert';
 import chalk from 'chalk';
+import crypto from 'crypto';
+import fs from 'fs';
 import { TransformInputOptions } from 'metro';
 import baseJSBundle from 'metro/src/DeltaBundler/Serializers/baseJSBundle';
 import {
@@ -22,7 +24,10 @@ import { TransformProfile } from 'metro-babel-transformer';
 import type { CustomResolverOptions } from 'metro-resolver/src/types';
 import path from 'path';
 
-import { createServerComponentsMiddleware } from './createServerComponentsMiddleware';
+import {
+  createServerComponentsMiddleware,
+  fileURLToFilePath,
+} from './createServerComponentsMiddleware';
 import { createRouteHandlerMiddleware } from './createServerRouteMiddleware';
 import { ExpoRouterServerManifestV1, fetchManifest } from './fetchRouterManifest';
 import { instantiateMetroAsync } from './instantiateMetro';
@@ -40,6 +45,7 @@ import { observeAnyFileChanges, observeFileChanges } from './waitForMetroToObser
 import { BundleAssetWithFileHashes, ExportAssetMap } from '../../../export/saveAssets';
 import { Log } from '../../../log';
 import getDevClientProperties from '../../../utils/analytics/getDevClientProperties';
+import { fileExistsAsync } from '../../../utils/dir';
 import { env } from '../../../utils/env';
 import { CommandError } from '../../../utils/errors';
 import { getFreePortAsync } from '../../../utils/port';
@@ -53,6 +59,10 @@ import {
 import { ContextModuleSourceMapsMiddleware } from '../middleware/ContextModuleSourceMapsMiddleware';
 import { CreateFileMiddleware } from '../middleware/CreateFileMiddleware';
 import { DevToolsPluginMiddleware } from '../middleware/DevToolsPluginMiddleware';
+import {
+  createDomComponentsMiddleware,
+  getDomComponentVirtualProxy,
+} from '../middleware/DomComponentsMiddleware';
 import { FaviconMiddleware } from '../middleware/FaviconMiddleware';
 import { HistoryFallbackMiddleware } from '../middleware/HistoryFallbackMiddleware';
 import { InterstitialPageMiddleware } from '../middleware/InterstitialPageMiddleware';
@@ -561,7 +571,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
   async legacySinglePageExportBundleAsync(
     options: Omit<
       ExpoMetroOptions,
-      'baseUrl' | 'routerRoot' | 'asyncRoutes' | 'isExporting' | 'serializerOutput' | 'environment'
+      'routerRoot' | 'asyncRoutes' | 'isExporting' | 'serializerOutput' | 'environment'
     >,
     extraOptions: {
       sourceMapUrl?: string;
@@ -569,6 +579,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     } = {}
   ): Promise<{ artifacts: SerialAsset[]; assets: readonly BundleAssetWithFileHashes[] }> {
     const { baseUrl, routerRoot, isExporting } = this.instanceMetroOptions;
+    assert(options.mainModuleName != null, 'mainModuleName must be provided in options.');
     assert(
       baseUrl != null && routerRoot != null && isExporting != null,
       'The server must be started before calling legacySinglePageExportBundleAsync.'
@@ -628,6 +639,30 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     );
   }
 
+  async getDomComponentVirtualEntryModuleAsync(file: string) {
+    const filePath = file.startsWith('file://') ? fileURLToFilePath(file) : file;
+
+    const hash = crypto.createHash('sha1').update(filePath).digest('hex');
+
+    const generatedEntry = path.join(this.projectRoot, '.expo/@dom', hash + '.js');
+
+    const entryFile = getDomComponentVirtualProxy(generatedEntry, filePath);
+
+    fs.mkdirSync(path.dirname(entryFile.filePath), { recursive: true });
+
+    const exists = await fileExistsAsync(entryFile.filePath);
+    // TODO: Assert no default export at runtime.
+    await fs.promises.writeFile(entryFile.filePath, entryFile.contents);
+
+    if (!exists) {
+      // Give time for watchman to compute the file...
+      // TODO: Virtual modules which can have dependencies.
+      await new Promise((res) => setTimeout(res, 1000));
+    }
+
+    return generatedEntry;
+  }
+
   protected async startImplementationAsync(
     options: BundlerStartOptions
   ): Promise<DevServerInstance> {
@@ -648,7 +683,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     const appDir = path.join(this.projectRoot, routerRoot);
     const mode = options.mode ?? 'development';
 
-    this.instanceMetroOptions = {
+    const instanceMetroOptions = {
       isExporting: !!options.isExporting,
       baseUrl,
       mode,
@@ -658,6 +693,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       asyncRoutes,
       // Options that are changing between platforms like engine, platform, and environment aren't set here.
     };
+    this.instanceMetroOptions = instanceMetroOptions;
 
     const parsedOptions = {
       port: options.port,
@@ -715,6 +751,21 @@ export class MetroBundlerDevServer extends BundlerDevServer {
         },
       });
       middleware.use(deepLinkMiddleware.getHandler());
+
+      const serverRoot = getMetroServerRoot(this.projectRoot);
+
+      // Add support for DOM components.
+      // TODO: Maybe put behind a flag for now?
+      middleware.use(
+        createDomComponentsMiddleware(
+          {
+            projectRoot: this.projectRoot,
+            metroRoot: serverRoot,
+            getDevServerUrl: this.getDevServerUrlOrAssert.bind(this),
+          },
+          instanceMetroOptions
+        )
+      );
 
       middleware.use(new CreateFileMiddleware(this.projectRoot).getHandler());